Simplify app layout and add option to delete apps
