--- packages/builder/src/components/start/Steps/index.js
@@ -1,2 +1,0 @@
-export { default as Info } from "./Info.svelte"
-export { default as User } from "./User.svelte"
